refactor(cmdb): 提取模型枚举字段策略
关联 Issue #3378：ModelManage 混合承载枚举字段规则与模型管理职责。

变更内容：

- 新增 ModelAttributePolicy 承载枚举字段策略

- 保留 ModelManage 同名兼容入口与签名

- 增加抽取与默认值行为契约测试
