+++ server/apps/cmdb/services/model.py
@@ -45,6 +45,7 @@ from apps.cmdb.services.auto_relation_rule import (
     validate_auto_relation_rule_set_payload,
 )
 from apps.cmdb.services.classification import ClassificationManage
+from apps.cmdb.services.model_attribute_policy import ModelAttributePolicy
 from apps.cmdb.services.model_visibility import BusinessModelVisibility
 from apps.cmdb.services.unique_rule import (
     UniqueRulePayload,
@@ -150,49 +151,13 @@ class ModelManage(object):
     def _clone_options(option: list[dict]) -> list[dict]:
         return [dict(item) for item in option]
 
-    @staticmethod
-    def _normalize_default_value(raw_value: Any) -> list[str]:
-        if raw_value in (None, ""):
-            return []
-
-        source = raw_value if isinstance(raw_value, list) else [raw_value]
-        seen: set[str] = set()
-        normalized: list[str] = []
-        for item in source:
-            value = str(item or "").strip()
-            if not value or value in seen:
-                continue
-            seen.add(value)
-            normalized.append(value)
-        return normalized
-
-    @staticmethod
-    def sanitize_attr_default_value(attr: dict, log_context: str = "") -> dict:
-        item = dict(attr)
-        normalized = ModelManage._normalize_default_value(item.get("default_value"))
-        if item.get("attr_type") != "enum":
-            item["default_value"] = normalized
-            return item
-
-        runtime_options = ModelManage.resolve_runtime_enum_options(item)
-        valid_ids = {str(option.get("id")).strip() for option in runtime_options if isinstance(option, dict) and str(option.get("id", "")).strip()}
-        sanitized = [value for value in normalized if value in valid_ids]
-        select_mode = item.get("enum_select_mode", ENUM_SELECT_MODE_DEFAULT)
-        if select_mode != "multiple":
-            sanitized = sanitized[:1]
-
-        removed_values = [value for value in normalized if value not in sanitized]
-        if removed_values:
-            logger.info(
-                "[ModelDefaultValue] pruned stale defaults context=%s attr_id=%s removed=%s rule_type=%s",
-                log_context or "runtime",
-                item.get("attr_id"),
-                removed_values,
-                item.get("enum_rule_type", "custom"),
-            )
-
-        item["default_value"] = sanitized
-        return item
+    _normalize_default_value = staticmethod(ModelAttributePolicy._normalize_default_value)
+    sanitize_attr_default_value = staticmethod(ModelAttributePolicy.sanitize_attr_default_value)
+    normalize_enum_public_binding = staticmethod(ModelAttributePolicy.normalize_enum_public_binding)
+    validate_enum_rule_immutable = staticmethod(ModelAttributePolicy.validate_enum_rule_immutable)
+    ensure_enum_select_mode = staticmethod(ModelAttributePolicy.ensure_enum_select_mode)
+    validate_enum_select_mode_immutable = staticmethod(ModelAttributePolicy.validate_enum_select_mode_immutable)
+    resolve_runtime_enum_options = staticmethod(ModelAttributePolicy.resolve_runtime_enum_options)
 
     @staticmethod
     def _normalize_attr_constraints(attrs: list[dict]) -> list[dict]:
@@ -306,184 +271,6 @@ class ModelManage(object):
     def _guard_protected_model_attr(attr_id: str, action: str):
         if attr_id in PROTECTED_MODEL_ATTR_IDS:
             raise BaseAppException(f"模型字段 {attr_id} 为系统字段，不允许{action}")
-
-    @staticmethod
-    def normalize_enum_public_binding(attr_info: dict, current_attr: dict | None = None) -> dict:
-        """
-        规范化 enum 公共选项库绑定信息并回填 option 快照。
-
-        处理逻辑：
-        1. 若 attr_type != enum，直接返回原 attr_info，不做处理。
-        2. enum_rule_type 默认为 custom。
-        3. 若 enum_rule_type == public_library：
-           - 校验 public_library_id 必填
-           - 从公共库拉取最新 options 并写入 option 字段（快照）
-        4. 若 enum_rule_type == custom：
-           - 清空 public_library_id
-           - 保持 option 不变
-
-        Args:
-            attr_info: 前端传入的属性配置
-            current_attr: 当前已存在的属性（更新场景下传入）
-
-        Returns:
-            dict: 规范化后的 attr_info（原地修改并返回）
-
-        Raises:
-            BaseAppException: 公共库不存在或配置不合法时抛出
-        """
-        if attr_info.get("attr_type") != "enum":
-            return attr_info
-
-        option_value = attr_info.get("option")
-        if isinstance(option_value, dict) and option_value.get("enum_rule_type"):
-            attr_info["enum_rule_type"] = option_value.get("enum_rule_type", "custom")
-            attr_info["public_library_id"] = option_value.get("public_library_id")
-            if "enum_select_mode" in option_value:
-                attr_info["enum_select_mode"] = option_value.get("enum_select_mode")
-            attr_info["option"] = option_value.get("option", [])
-
-        enum_rule_type = attr_info.get("enum_rule_type", "custom")
-        attr_info["enum_rule_type"] = enum_rule_type
-
-        if enum_rule_type == "public_library":
-            public_library_id = attr_info.get("public_library_id")
-            if not public_library_id:
-                raise BaseAppException("绑定公共选项库时 public_library_id 必填")
-
-            from apps.cmdb.services.public_enum_library import get_library_or_raise
-
-            library = get_library_or_raise(public_library_id)
-            attr_info["option"] = library.options
-            attr_info["public_library_id"] = public_library_id
-
-            logger.info(
-                f"[EnumPublicBinding] normalized attr_id={attr_info.get('attr_id')}, "
-                f"enum_rule_type={enum_rule_type}, public_library_id={public_library_id}"
-            )
-        else:
-            attr_info["enum_rule_type"] = "custom"
-            attr_info["public_library_id"] = None
-
-        return attr_info
-
-    @staticmethod
-    def validate_enum_rule_immutable(current_attr: dict, incoming_attr: dict) -> None:
-        """
-        校验字段创建后 enum_rule_type 不可切换。
-
-        规则：
-        - 仅对 attr_type == enum 的字段生效
-        - 创建时：任意 enum_rule_type（custom / public_library）均可
-        - 更新时：不允许从 custom 切换到 public_library，或反之
-
-        Args:
-            current_attr: 当前已存储的属性定义
-            incoming_attr: 本次请求传入的属性定义
-
-        Raises:
-            BaseAppException: 规则类型切换时抛出
-        """
-        if current_attr.get("attr_type") != "enum":
-            return
-        if incoming_attr.get("attr_type") != "enum":
-            return
-
-        current_rule = current_attr.get("enum_rule_type", "custom")
-        incoming_rule = incoming_attr.get("enum_rule_type", "custom")
-
-        if current_rule != incoming_rule:
-            raise BaseAppException(f"枚举字段创建后规则类型不可切换（当前: {current_rule}）")
-
-    @staticmethod
-    def ensure_enum_select_mode(attr_info: dict) -> dict:
-        """
-        确保枚举字段具有 enum_select_mode 属性。
-
-        规则：
-        - 仅对 attr_type == enum 的字段生效
-        - 若未提供 enum_select_mode，则默认设置为 single
-
-        Args:
-            attr_info: 属性配置字典
-
-        Returns:
-            dict: 规范化后的 attr_info（原地修改并返回）
-        """
-        if attr_info.get("attr_type") != "enum":
-            return attr_info
-
-        if "enum_select_mode" not in attr_info:
-            attr_info["enum_select_mode"] = ENUM_SELECT_MODE_DEFAULT
-
-        return attr_info
-
-    @staticmethod
-    def validate_enum_select_mode_immutable(current_attr: dict, incoming_attr: dict) -> None:
-        """
-        校验字段创建后 enum_select_mode 不可切换。
-
-        规则：
-        - 仅对 attr_type == enum 的字段生效
-        - 创建时：任意 enum_select_mode（single / multiple）均可
-        - 更新时：不允许从 single 切换到 multiple，或反之
-
-        Args:
-            current_attr: 当前已存储的属性定义
-            incoming_attr: 本次请求传入的属性定义
-
-        Raises:
-            BaseAppException: 选择模式切换时抛出
-        """
-        if current_attr.get("attr_type") != "enum":
-            return
-        if incoming_attr.get("attr_type") != "enum":
-            return
-
-        current_mode = current_attr.get("enum_select_mode", ENUM_SELECT_MODE_DEFAULT)
-        incoming_mode = incoming_attr.get("enum_select_mode", current_mode)
-
-        if current_mode != incoming_mode:
-            raise BaseAppException(f"枚举字段创建后选择模式不可切换（当前: {current_mode}）")
-
-    @staticmethod
-    def resolve_runtime_enum_options(attr: dict) -> list[dict]:
-        """
-        运行时解析枚举选项。
-
-        口径：
-        - 若 enum_rule_type == public_library，优先从公共库实时拉取
-        - 若公共库不存在或查询失败，回退到字段快照（attr.option）
-        - 若 enum_rule_type == custom，直接返回 attr.option
-
-        Args:
-            attr: 字段属性定义
-
-        Returns:
-            list[dict]: 枚举选项列表 [{"id": str, "name": str}, ...]
-        """
-        if attr.get("attr_type") != "enum":
-            return []
-
-        enum_rule_type = attr.get("enum_rule_type", "custom")
-        option = attr.get("option", [])
-
-        if enum_rule_type != "public_library":
-            return option if isinstance(option, list) else []
-
-        public_library_id = str(attr.get("public_library_id") or "").strip()
-        if not public_library_id:
-            return option if isinstance(option, list) else []
-
-        try:
-            from apps.cmdb.services.public_enum_library import get_library_or_raise
-
-            library = get_library_or_raise(public_library_id)
-            runtime_options = library.options
-            return runtime_options if isinstance(runtime_options, list) else []
-        except Exception as e:
-            logger.warning(f"[EnumPublicBinding] resolve_runtime_enum_options fallback to snapshot, public_library_id={public_library_id}, error={e}")
-            return option if isinstance(option, list) else []
 
     @staticmethod
     def _add_display_field_to_attrs(attrs: list, attr_info: dict, model_id: str, is_pre: bool = False):
